fix(access): build the route map from a leaf module
routePermissions reads the permission groups at module scope, and
permissions.ts sits inside the router import cycle (router ->
routePermissions -> permissions -> services/user -> services/api ->
router). In the browser that threw 'Cannot access CHAT_VIEW_PERMISSIONS
before initialization' and took the sidebar down with it.

The unit tests could not see it: they mock services/user, which breaks
the cycle. Caught by loading the actual page.

The groups now live in a leaf module that imports nothing. Only
hasAnyPermission still comes from permissions.ts, and that is called at
runtime, where a partially-initialised module is harmless.
